requirements.txt
# Python dependencies for Vercel serverless functions
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
reportlab==4.0.7
PyPDF2==3.0.1
beautifulsoup4==4.12.2
